+++ flickrj/src/com/aetrion/flickr/commons/Institution.java
@@ -10,68 +10,69 @@ import java.util.Date;
  */
 public class Institution {
 	private static final long serialVersionUID = 12L;
-    String id;
-    String name;
-    Date dateLaunch;
-    String siteUrl;
-    String licenseUrl;
-    String flickrUrl;
+	String id;
+	String name;
+	Date dateLaunch;
+	String siteUrl;
+	String licenseUrl;
+	String flickrUrl;
 
-    public String getId() {
-        return id;
-    }
+	public String getId() {
+		return id;
+	}
 
-    public void setId(String id) {
-        this.id = id;
-    }
+	public void setId(String id) {
+		this.id = id;
+	}
 
-    public String getName() {
-        return name;
-    }
+	public String getName() {
+		return name;
+	}
 
-    public void setName(String name) {
-        this.name = name;
-    }
+	public void setName(String name) {
+		this.name = name;
+	}
 
-    public Date getDateLaunch() {
-        return dateLaunch;
-    }
+	public Date getDateLaunch() {
+		return dateLaunch;
+	}
 
-    public void setDateLaunch(Date dateLaunch) {
-        this.dateLaunch = dateLaunch;
-    }
+	public void setDateLaunch(Date dateLaunch) {
+		this.dateLaunch = dateLaunch;
+	}
 
-    public void setDateLaunch(long date) {
-        setDateLaunch(new Date(date));
-    }
+	public void setDateLaunch(long date) {
+		setDateLaunch(new Date(date));
+	}
 
-    public void setDateLaunch(String date) {
-        if (date == null || "".equals(date)) return;
-        setDateLaunch(Long.parseLong(date) * (long) 1000);
-    }
+	public void setDateLaunch(String date) {
+		if (date == null || "".equals(date))
+			return;
+		setDateLaunch(Long.parseLong(date) * 1000);
+	}
 
-    public String getSiteUrl() {
-        return siteUrl;
-    }
+	public String getSiteUrl() {
+		return siteUrl;
+	}
 
-    public void setSiteUrl(String siteUrl) {
-        this.siteUrl = siteUrl;
-    }
+	public void setSiteUrl(String siteUrl) {
+		this.siteUrl = siteUrl;
+	}
 
-    public String getLicenseUrl() {
-        return licenseUrl;
-    }
+	public String getLicenseUrl() {
+		return licenseUrl;
+	}
 
-    public void setLicenseUrl(String licenseUrl) {
-        this.licenseUrl = licenseUrl;
-    }
+	public void setLicenseUrl(String licenseUrl) {
+		this.licenseUrl = licenseUrl;
+	}
 
-    public String getFlickrUrl() {
-        return flickrUrl;
-    }
+	public String getFlickrUrl() {
+		return flickrUrl;
+	}
 
-    public void setFlickrUrl(String flickrUrl) {
-        this.flickrUrl = flickrUrl;
-    }
+	public void setFlickrUrl(String flickrUrl) {
+		this.flickrUrl = flickrUrl;
+	}
 
 }
